bash toolset: move argv danger detection into dangerous_args.py
Extract the per-command argv checks (find/sort/uniq rules, the getopt-style
parser, the abbreviation matcher, the ARGV_CHECKERS registry / dispatch, and
the shared benign-redirect-target helper) into a new dangerous_args.py.

validation.py keeps the AST extractor, the ValidationResult machinery, and
check_dangerous_argv / validate_command, importing the detection API
(dangerous_argv_reason, ARGV_CHECKERS, is_benign_redirect_target). Pure move
plus rename of the now cross-module names; behavior unchanged.
